Added the entity for beer closes #6

+++ src/main/java/abhash/springframework/web/model/BeerDto.java
@@ -21,7 +21,7 @@ public class BeerDto {
 	private long price;
 	
 	private UUID uuid;
-	private int version;
+	private Long version;
 	
 	private OffsetDateTime createdDate;
 	private OffsetDateTime lastModifiedDate;
